+++ packages/ethereum-contracts/contracts/utils/SuperfluidFrameworkDeploymentSteps.t.sol
@@ -194,6 +194,16 @@ contract SuperfluidFrameworkDeploymentSteps {
                 gdaV1Logic.superfluidPoolBeacon().upgradeTo(address(superfluidPoolLogic));
                 gdaV1Logic.superfluidPoolBeacon().transferOwnership(address(host));
             }
+
+            bytes32 aclPoolConnectExclusiveRoleAdmin = keccak256("ACL_POOL_CONNECT_EXCLUSIVE_ROLE_ADMIN");
+            SimpleACL(address(host.getSimpleACL())).setRoleAdmin(
+                gdaV1.ACL_POOL_CONNECT_EXCLUSIVE_ROLE(),
+                aclPoolConnectExclusiveRoleAdmin
+            );
+            SimpleACL(address(host.getSimpleACL())).grantRole(
+                aclPoolConnectExclusiveRoleAdmin,
+                address(gdaV1)
+            );
         } else if (step == 3) {// PERIPHERAL CONTRACTS: NFT Proxy and Logic
             {
                 poolAdminNFT = PoolAdminNFT(address(ProxyDeployerLibrary.deployUUPSProxy()));