+++ packages/ethereum-contracts/hardhat.config.ts
@@ -99,7 +99,7 @@ const config: HardhatUserConfig = {
                 enabled: true,
                 runs: 200,
             },
-            evmVersion: "paris",
+            evmVersion: "shanghai",
         },
     },
     paths: {